% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_functions.R
\docType{package}
\name{content_googleAuthR}
\alias{content_googleAuthR}
\alias{content_googleAuthR-package}
\title{Content API for Shopping
Manages product items, inventory, and Merchant Center accounts for Google Shopping.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:40:20
filename: /Users/mark/dev/R/autoGoogleAPI/googlecontentv2sandbox.auto/R/content_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/content
}
}
