% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{Error}
\alias{Error}
\title{Content API for Shopping Objects 
Manages product items, inventory, and Merchant Center accounts for Google Shopping.}
\usage{
Error(domain = NULL, message = NULL, reason = NULL)
}
\arguments{
\item{domain}{The domain of the error}

\item{message}{A description of the error}

\item{reason}{The error code}
}
\value{
Error object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:40:20
filename: /Users/mark/dev/R/autoGoogleAPI/googlecontentv2sandbox.auto/R/content_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Error Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An error returned by the API.
}
