% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{Errors}
\alias{Errors}
\title{Errors Object}
\usage{
Errors(code = NULL, errors = NULL, message = NULL)
}
\arguments{
\item{code}{The HTTP status of the first error in errors}

\item{errors}{A list of errors}

\item{message}{The message of the first error in errors}
}
\value{
Errors object
}
\description{
Errors Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of errors returned by a failed batch entry.
}
