% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrderLineItem}
\alias{OrderLineItem}
\title{OrderLineItem Object}
\usage{
OrderLineItem(cancellations = NULL, id = NULL, price = NULL,
  product = NULL, quantityCanceled = NULL, quantityDelivered = NULL,
  quantityOrdered = NULL, quantityPending = NULL, quantityReturned = NULL,
  quantityShipped = NULL, returnInfo = NULL, returns = NULL,
  shippingDetails = NULL, tax = NULL)
}
\arguments{
\item{cancellations}{Cancellations of the line item}

\item{id}{The id of the line item}

\item{price}{Total price for the line item}

\item{product}{Product data from the time of the order placement}

\item{quantityCanceled}{Number of items canceled}

\item{quantityDelivered}{Number of items delivered}

\item{quantityOrdered}{Number of items ordered}

\item{quantityPending}{Number of items pending}

\item{quantityReturned}{Number of items returned}

\item{quantityShipped}{Number of items shipped}

\item{returnInfo}{Details of the return policy for the line item}

\item{returns}{Returns of the line item}

\item{shippingDetails}{Details of the requested shipping for the line item}

\item{tax}{Total tax amount for the line item}
}
\value{
OrderLineItem object
}
\description{
OrderLineItem Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

