% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrderLineItemProductVariantAttribute}
\alias{OrderLineItemProductVariantAttribute}
\title{OrderLineItemProductVariantAttribute Object}
\usage{
OrderLineItemProductVariantAttribute(dimension = NULL, value = NULL)
}
\arguments{
\item{dimension}{The dimension of the variant}

\item{value}{The value for the dimension}
}
\value{
OrderLineItemProductVariantAttribute object
}
\description{
OrderLineItemProductVariantAttribute Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

