% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrderLineItemShippingDetailsMethod}
\alias{OrderLineItemShippingDetailsMethod}
\title{OrderLineItemShippingDetailsMethod Object}
\usage{
OrderLineItemShippingDetailsMethod(carrier = NULL, maxDaysInTransit = NULL,
  methodName = NULL, minDaysInTransit = NULL)
}
\arguments{
\item{carrier}{The carrier for the shipping}

\item{maxDaysInTransit}{Maximum transit time}

\item{methodName}{The name of the shipping method}

\item{minDaysInTransit}{Minimum transit time}
}
\value{
OrderLineItemShippingDetailsMethod object
}
\description{
OrderLineItemShippingDetailsMethod Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

