% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrderPromotion}
\alias{OrderPromotion}
\title{OrderPromotion Object}
\usage{
OrderPromotion(benefits = NULL, effectiveDates = NULL,
  genericRedemptionCode = NULL, id = NULL, longTitle = NULL,
  productApplicability = NULL, redemptionChannel = NULL)
}
\arguments{
\item{benefits}{No description}

\item{effectiveDates}{The date and time frame when the promotion is active and ready for validation review}

\item{genericRedemptionCode}{Optional}

\item{id}{The unique ID of the promotion}

\item{longTitle}{The full title of the promotion}

\item{productApplicability}{Whether the promotion is applicable to all products or only specific products}

\item{redemptionChannel}{Indicates that the promotion is valid online}
}
\value{
OrderPromotion object
}
\description{
OrderPromotion Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
