% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrderPromotionBenefit}
\alias{OrderPromotionBenefit}
\title{OrderPromotionBenefit Object}
\usage{
OrderPromotionBenefit(discount = NULL, offerIds = NULL, subType = NULL,
  taxImpact = NULL, type = NULL)
}
\arguments{
\item{discount}{The discount in the order price when the promotion is applied}

\item{offerIds}{The OfferId(s) that were purchased in this order and map to this specific benefit of the promotion}

\item{subType}{Further describes the benefit of the promotion}

\item{taxImpact}{The impact on tax when the promotion is applied}

\item{type}{Describes whether the promotion applies to products (e}
}
\value{
OrderPromotionBenefit object
}
\description{
OrderPromotionBenefit Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
