% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrderReturn}
\alias{OrderReturn}
\title{OrderReturn Object}
\usage{
OrderReturn(actor = NULL, creationDate = NULL, quantity = NULL,
  reason = NULL, reasonText = NULL)
}
\arguments{
\item{actor}{The actor that created the refund}

\item{creationDate}{Date on which the item has been created, in ISO 8601 format}

\item{quantity}{Quantity that is returned}

\item{reason}{The reason for the return}

\item{reasonText}{The explanation of the reason}
}
\value{
OrderReturn object
}
\description{
OrderReturn Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
