% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_functions.R
\name{orders.list}
\alias{orders.list}
\title{Lists the orders in your Merchant Center account.}
\usage{
orders.list(merchantId, acknowledged = NULL, maxResults = NULL,
  orderBy = NULL, pageToken = NULL, placedDateEnd = NULL,
  placedDateStart = NULL, statuses = NULL)
}
\arguments{
\item{merchantId}{The ID of the managing account}

\item{acknowledged}{Obtains orders that match the acknowledgement status}

\item{maxResults}{The maximum number of orders to return in the response, used for paging}

\item{orderBy}{The ordering of the returned list}

\item{pageToken}{The token returned by the previous request}

\item{placedDateEnd}{Obtains orders placed before this date (exclusively), in ISO 8601 format}

\item{placedDateStart}{Obtains orders placed after this date (inclusively), in ISO 8601 format}

\item{statuses}{Obtains orders that match any of the specified statuses}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/content
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/content)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/shopping-content}{Google Documentation}
}

