% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_functions.R
\name{orders.refund}
\alias{orders.refund}
\title{Refund a portion of the order, up to the full amount paid. This method can only be called for non-multi-client accounts.}
\usage{
orders.refund(OrdersRefundRequest, merchantId, orderId)
}
\arguments{
\item{OrdersRefundRequest}{The \link{OrdersRefundRequest} object to pass to this method}

\item{merchantId}{The ID of the managing account}

\item{orderId}{The ID of the order to refund}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/content
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/content)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/shopping-content}{Google Documentation}

Other OrdersRefundRequest functions: \code{\link{OrdersRefundRequest}}
}
