% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersAcknowledgeRequest}
\alias{OrdersAcknowledgeRequest}
\title{OrdersAcknowledgeRequest Object}
\usage{
OrdersAcknowledgeRequest(operationId = NULL)
}
\arguments{
\item{operationId}{The ID of the operation}
}
\value{
OrdersAcknowledgeRequest object
}
\description{
OrdersAcknowledgeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other OrdersAcknowledgeRequest functions: \code{\link{orders.acknowledge}}
}

