% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersCancelRequest}
\alias{OrdersCancelRequest}
\title{OrdersCancelRequest Object}
\usage{
OrdersCancelRequest(operationId = NULL, reason = NULL, reasonText = NULL)
}
\arguments{
\item{operationId}{The ID of the operation}

\item{reason}{The reason for the cancellation}

\item{reasonText}{The explanation of the reason}
}
\value{
OrdersCancelRequest object
}
\description{
OrdersCancelRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other OrdersCancelRequest functions: \code{\link{orders.cancel}}
}
