% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersCancelResponse}
\alias{OrdersCancelResponse}
\title{OrdersCancelResponse Object}
\usage{
OrdersCancelResponse(executionStatus = NULL)
}
\arguments{
\item{executionStatus}{The status of the execution}
}
\value{
OrdersCancelResponse object
}
\description{
OrdersCancelResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
