% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersCreateTestOrderRequest}
\alias{OrdersCreateTestOrderRequest}
\title{OrdersCreateTestOrderRequest Object}
\usage{
OrdersCreateTestOrderRequest(templateName = NULL, testOrder = NULL)
}
\arguments{
\item{templateName}{The test order template to use}

\item{testOrder}{The test order to create}
}
\value{
OrdersCreateTestOrderRequest object
}
\description{
OrdersCreateTestOrderRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other OrdersCreateTestOrderRequest functions: \code{\link{orders.createtestorder}}
}

