% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersCreateTestOrderResponse}
\alias{OrdersCreateTestOrderResponse}
\title{OrdersCreateTestOrderResponse Object}
\usage{
OrdersCreateTestOrderResponse(orderId = NULL)
}
\arguments{
\item{orderId}{The ID of the newly created test order}
}
\value{
OrdersCreateTestOrderResponse object
}
\description{
OrdersCreateTestOrderResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

