% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersCustomBatchRequestEntryCancel}
\alias{OrdersCustomBatchRequestEntryCancel}
\title{OrdersCustomBatchRequestEntryCancel Object}
\usage{
OrdersCustomBatchRequestEntryCancel(reason = NULL, reasonText = NULL)
}
\arguments{
\item{reason}{The reason for the cancellation}

\item{reasonText}{The explanation of the reason}
}
\value{
OrdersCustomBatchRequestEntryCancel object
}
\description{
OrdersCustomBatchRequestEntryCancel Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

