% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersCustomBatchRequestEntryCancelLineItem}
\alias{OrdersCustomBatchRequestEntryCancelLineItem}
\title{OrdersCustomBatchRequestEntryCancelLineItem Object}
\usage{
OrdersCustomBatchRequestEntryCancelLineItem(amount = NULL,
  lineItemId = NULL, quantity = NULL, reason = NULL, reasonText = NULL)
}
\arguments{
\item{amount}{Amount to refund for the cancelation}

\item{lineItemId}{The ID of the line item to cancel}

\item{quantity}{The quantity to cancel}

\item{reason}{The reason for the cancellation}

\item{reasonText}{The explanation of the reason}
}
\value{
OrdersCustomBatchRequestEntryCancelLineItem object
}
\description{
OrdersCustomBatchRequestEntryCancelLineItem Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
