% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersCustomBatchRequestEntryReturnLineItem}
\alias{OrdersCustomBatchRequestEntryReturnLineItem}
\title{OrdersCustomBatchRequestEntryReturnLineItem Object}
\usage{
OrdersCustomBatchRequestEntryReturnLineItem(lineItemId = NULL,
  quantity = NULL, reason = NULL, reasonText = NULL)
}
\arguments{
\item{lineItemId}{The ID of the line item to return}

\item{quantity}{The quantity to return}

\item{reason}{The reason for the return}

\item{reasonText}{The explanation of the reason}
}
\value{
OrdersCustomBatchRequestEntryReturnLineItem object
}
\description{
OrdersCustomBatchRequestEntryReturnLineItem Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

