% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersCustomBatchRequestEntryUpdateShipment}
\alias{OrdersCustomBatchRequestEntryUpdateShipment}
\title{OrdersCustomBatchRequestEntryUpdateShipment Object}
\usage{
OrdersCustomBatchRequestEntryUpdateShipment(carrier = NULL,
  shipmentId = NULL, status = NULL, trackingId = NULL)
}
\arguments{
\item{carrier}{The carrier handling the shipment}

\item{shipmentId}{The ID of the shipment}

\item{status}{New status for the shipment}

\item{trackingId}{The tracking id for the shipment}
}
\value{
OrdersCustomBatchRequestEntryUpdateShipment object
}
\description{
OrdersCustomBatchRequestEntryUpdateShipment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

