% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersCustomBatchResponse}
\alias{OrdersCustomBatchResponse}
\title{OrdersCustomBatchResponse Object}
\usage{
OrdersCustomBatchResponse(entries = NULL)
}
\arguments{
\item{entries}{The result of the execution of the batch requests}
}
\value{
OrdersCustomBatchResponse object
}
\description{
OrdersCustomBatchResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

