% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersCustomBatchResponseEntry}
\alias{OrdersCustomBatchResponseEntry}
\title{OrdersCustomBatchResponseEntry Object}
\usage{
OrdersCustomBatchResponseEntry(batchId = NULL, errors = NULL,
  executionStatus = NULL, order = NULL)
}
\arguments{
\item{batchId}{The ID of the request entry this entry responds to}

\item{errors}{A list of errors defined if and only if the request failed}

\item{executionStatus}{The status of the execution}

\item{order}{The retrieved order}
}
\value{
OrdersCustomBatchResponseEntry object
}
\description{
OrdersCustomBatchResponseEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
