% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersGetTestOrderTemplateResponse}
\alias{OrdersGetTestOrderTemplateResponse}
\title{OrdersGetTestOrderTemplateResponse Object}
\usage{
OrdersGetTestOrderTemplateResponse(template = NULL)
}
\arguments{
\item{template}{The requested test order template}
}
\value{
OrdersGetTestOrderTemplateResponse object
}
\description{
OrdersGetTestOrderTemplateResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

