% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersShipLineItemsRequest}
\alias{OrdersShipLineItemsRequest}
\title{OrdersShipLineItemsRequest Object}
\usage{
OrdersShipLineItemsRequest(carrier = NULL, lineItems = NULL,
  operationId = NULL, shipmentId = NULL, trackingId = NULL)
}
\arguments{
\item{carrier}{The carrier handling the shipment}

\item{lineItems}{Line items to ship}

\item{operationId}{The ID of the operation}

\item{shipmentId}{The ID of the shipment}

\item{trackingId}{The tracking id for the shipment}
}
\value{
OrdersShipLineItemsRequest object
}
\description{
OrdersShipLineItemsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other OrdersShipLineItemsRequest functions: \code{\link{orders.shiplineitems}}
}

