% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersUpdateShipmentRequest}
\alias{OrdersUpdateShipmentRequest}
\title{OrdersUpdateShipmentRequest Object}
\usage{
OrdersUpdateShipmentRequest(carrier = NULL, operationId = NULL,
  shipmentId = NULL, status = NULL, trackingId = NULL)
}
\arguments{
\item{carrier}{The carrier handling the shipment}

\item{operationId}{The ID of the operation}

\item{shipmentId}{The ID of the shipment}

\item{status}{New status for the shipment}

\item{trackingId}{The tracking id for the shipment}
}
\value{
OrdersUpdateShipmentRequest object
}
\description{
OrdersUpdateShipmentRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other OrdersUpdateShipmentRequest functions: \code{\link{orders.updateshipment}}
}

