% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersUpdateShipmentResponse}
\alias{OrdersUpdateShipmentResponse}
\title{OrdersUpdateShipmentResponse Object}
\usage{
OrdersUpdateShipmentResponse(executionStatus = NULL)
}
\arguments{
\item{executionStatus}{The status of the execution}
}
\value{
OrdersUpdateShipmentResponse object
}
\description{
OrdersUpdateShipmentResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

