% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{TestOrderLineItem}
\alias{TestOrderLineItem}
\title{TestOrderLineItem Object}
\usage{
TestOrderLineItem(product = NULL, quantityOrdered = NULL,
  returnInfo = NULL, shippingDetails = NULL, unitTax = NULL)
}
\arguments{
\item{product}{Product data from the time of the order placement}

\item{quantityOrdered}{Number of items ordered}

\item{returnInfo}{Details of the return policy for the line item}

\item{shippingDetails}{Details of the requested shipping for the line item}

\item{unitTax}{Unit tax for the line item}
}
\value{
TestOrderLineItem object
}
\description{
TestOrderLineItem Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
