% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{TestOrderLineItemProduct}
\alias{TestOrderLineItemProduct}
\title{TestOrderLineItemProduct Object}
\usage{
TestOrderLineItemProduct(brand = NULL, channel = NULL, condition = NULL,
  contentLanguage = NULL, gtin = NULL, imageLink = NULL,
  itemGroupId = NULL, mpn = NULL, offerId = NULL, price = NULL,
  targetCountry = NULL, title = NULL, variantAttributes = NULL)
}
\arguments{
\item{brand}{Brand of the item}

\item{channel}{The item's channel}

\item{condition}{Condition or state of the item}

\item{contentLanguage}{The two-letter ISO 639-1 language code for the item}

\item{gtin}{Global Trade Item Number (GTIN) of the item}

\item{imageLink}{URL of an image of the item}

\item{itemGroupId}{Shared identifier for all variants of the same product}

\item{mpn}{Manufacturer Part Number (MPN) of the item}

\item{offerId}{An identifier of the item}

\item{price}{The price for the product}

\item{targetCountry}{The CLDR territory code of the target country of the product}

\item{title}{The title of the product}

\item{variantAttributes}{Variant attributes for the item}
}
\value{
TestOrderLineItemProduct object
}
\description{
TestOrderLineItemProduct Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

