% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{TestOrderPaymentMethod}
\alias{TestOrderPaymentMethod}
\title{TestOrderPaymentMethod Object}
\usage{
TestOrderPaymentMethod(expirationMonth = NULL, expirationYear = NULL,
  lastFourDigits = NULL, predefinedBillingAddress = NULL, type = NULL)
}
\arguments{
\item{expirationMonth}{The card expiration month (January = 1, February = 2 etc}

\item{expirationYear}{The card expiration year (4-digit, e}

\item{lastFourDigits}{The last four digits of the card number}

\item{predefinedBillingAddress}{The billing address}

\item{type}{The type of instrument}
}
\value{
TestOrderPaymentMethod object
}
\description{
TestOrderPaymentMethod Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

